Move patch-protobuf-root
