+++ packages/test/tsconfig.json
@@ -6,6 +6,7 @@
     "resolveJsonModule": true
   },
   "references": [
+    { "path": "../proto" },
     { "path": "../internal-workflow-common" },
     { "path": "../internal-non-workflow-common" },
     { "path": "../activity" },
